requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==4.7
scikit-learn
dill
pyaml
mlflow==2.12.2 
dagshub
fastapi
uvicorn
python-multipart

##-e .